+++ test/OpenTelemetry.Extensions.AWS.Tests/AWSXRayIdGeneratorTests.cs
@@ -1,4 +1,4 @@
-// <copyright file="TestAWSXRayIdGenerator.cs" company="OpenTelemetry Authors">
+// <copyright file="AWSXRayIdGeneratorTests.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,9 +19,9 @@ using System.Diagnostics;
 using OpenTelemetry.Trace;
 using Xunit;
 
-namespace OpenTelemetry.Contrib.Extensions.AWSXRay.Tests;
+namespace OpenTelemetry.Extensions.AWS.Tests;
 
-public class TestAWSXRayIdGenerator
+public class AWSXRayIdGeneratorTests
 {
     [Fact]
     public void TestGenerateTraceIdForRootNode()