+++ src/OpenTelemetry.Extensions.AWS/CHANGELOG.md
@@ -1,7 +1,10 @@
-# Changelog - OpenTelemetry.Contrib.Extensions.AWSXRay
+# Changelog - OpenTelemetry.Extensions.AWS
 
 ## Unreleased
 
+* Rename package from `OpenTelemetry.Contrib.Extensions.AWSXRay`
+  to `OpenTelemetry.Extensions.AWS`
+  ([#1232](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1232))
 * Updates to 1.5.1 of OpenTelemetry SDK.
   ([#1255](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1255))
 * Enhancement - AWSXRayIdGenerator - Generate X-Ray IDs with global Random
@@ -41,7 +44,7 @@ Released 2022-May-18
 Released 2021-Sep-20
 
 * Added AWS resource detectors ([#149](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/149))
-* Updated OpenTelemetry SDK package version to 1.1.0
+* Updated OTel SDK package version to 1.1.0
   ([#100](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/100))
 
 ## 1.0.1